python自动化办公/requirements.txt
appdirs==1.4.4
distlib==0.3.1
docx2pdf==0.1.7
filelock==3.0.12
importlib-metadata==1.7.0
lxml==4.6.2
numpy==1.19.5
pandas==1.2.0
pypiwin32==223
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
pywin32==227
scipy==1.6.0
six==1.15.0
smart-open==4.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
virtualenv==20.3.1
xlrd==1.2.0
zipp==3.4.0
